// src/modules/notification/types/notification-types.ts
export enum NotificationType {
  LIKE_POST = 'LIKE_POST',
  COMMENT_POST = 'COMMENT_POST',
  REPLY_COMMENT = 'REPLY_COMMENT',
  FOLLOW_USER = 'FOLLOW_USER',
  MESSAGE_SENT = 'MESSAGE_SENT',
  ORGANIZATION_JOINED = 'ORGANIZATION_JOINED',
  POST_PROMOTED_TO_GLOBAL = 'POST_PROMOTED_TO_GLOBAL',
  REPOST_POST = 'REPOST_POST',
  QUOTE_POST = 'QUOTE_POST',
}